Reconstruct early version of DECUS grep.c
This is a reconstruction of what I deduce to be the June 1983 revision
of DECUS grep, through comparative analysis of several versions
distributed by different sources.

The version distributed on comp.sources.unix [0] appears to have the
least changes, so I use it as the base. When it was submitted, a
moderator added a fix for BSD compatibility and annotated it; I revert
that. That change is its only unique difference compared to the other
recovered versions, so I am confident it is the closest to a common
version.

The earliest version in AROS contrib [1] is nearly identical, including
code formatting, except for some minor fixes.

The version bundled with early versions of the Zeus programmers editor
[2] has annotated changes for output compatibility with Zeus and adding
support for filename wildcards. With those removed, it is nearly
identical with a few changes that do not appear in other sources, so
were probably not annotated by the Zeus author.

The SNIPPETS collection bundled GREP.C, but it has many style changes.
It was first added in SNIP0492 [3], deleted, then reintroduced with a
completely different set of changes in SNIP9404 [4].

This grep.c is from comp.sources.unix without the BSD change. Relative
to it, all differences are unique to the other versions, no version a
subset of another. None of the changes fix bugs or (excluding Zeus
output changes) change behavior. Until I find more and older sources, I
am confident this represents a widely distributed version of DECUS grep.

[0]: https://sources.vsta.org/comp.sources.unix/volume3/decus_grep
[1]: https://github.com/aros-development-team/contrib/blob/3c9936f0f4a7f448617a7559b5e3d5525c2c331c/fish/grep/grep.c
[2]: https://archive.org/download/ZE32V215_ZIP/ZE32V215.ZIP/GREP.C_ (compressed)
[3]: http://annex.retroarchive.org/cdrom/cotc-1/PROGRAM/SNIP0492.ARJ
[4]: http://cd.textfiles.com/simtel/simtel/DISK1/DISC2/C/SNIP9404.ZIP