Combine “Pattern too complex” with `badpat` errors
“Pattern too complex” is the only error emitted by `compile`, which uses
`error` instead of `badpat`. I suspect it is because it concerns the
entire pattern, rather than a specific location; however, when debugging
it can be useful to know at what character the pattern becomes too
large. Combine it with the other pattern errors, but keep its
compatible-style formatting the same.