Fix line overrun for NCLASS
NCLASS will always match NUL, since it is not specifically excluded, so
it matches the line terminator and overruns the current line. The rest
of the pattern will then be matched against whatever follows the line
terminator. Up to 512 bytes, this will be the rest of `lbuf`. After
that, it is probably `pbuf`.